refactor: refactor notification cards
