+++ src/components/DropDown/DropDown.jsx
@@ -1,12 +1,19 @@
-import React from 'react';
+import React, {
+  useContext,
+  useCallback,
+  useState,
+  useEffect,
+} from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
+import { useSnackbar } from 'notistack';
 import {
   Collapse,
   Typography,
 } from '@material-ui/core';
-
-import Notifications from './Notifications';
+import useHttp from '../../hooks/http.hook';
+import AuthContext from '../../context/AuthContext';
+import NotificationCard from './NotificationCard';
 
 const useStyles = makeStyles((theme) => ({
   container: {
@@ -32,6 +39,97 @@ const useStyles = makeStyles((theme) => ({
 
 const DropDown = ({ isOpen }) => {
   const classes = useStyles();
+  const {
+    discardNotifications,
+    setNotification,
+    token,
+  } = useContext(AuthContext);
+  const { request, loading } = useHttp();
+  const { enqueueSnackbar } = useSnackbar();
+  const [nots, setNots] = useState([]);
+
+  const showSnackbar = useCallback((message, variant) => (
+    enqueueSnackbar(message, { variant })
+  ), [enqueueSnackbar]);
+
+  const getNots = async (tok, req) => {
+    const requestOptions = {
+      url: 'https://rsclone-back-end.herokuapp.com/api/auth/get_user',
+      method: 'GET',
+      headers: { Authorization: `Bearer ${tok}` },
+    };
+
+    const data = await req(requestOptions);
+
+    setNots(data.notifications || []);
+    localStorage.setItem('notifications', JSON.stringify({ notice: data.notifications }));
+  };
+
+  const comein = async (id, boardName) => {
+    try {
+      const requestOptions = {
+        url: `https://rsclone-back-end.herokuapp.com/api/board/comein/${id}`,
+        method: 'POST',
+        headers: { Authorization: `Bearer ${token}` },
+      };
+
+      await request(requestOptions);
+
+      await discardNotifications(id);
+      getNots(token, request);
+      showSnackbar(`Welcome to ${boardName}!`, 'success');
+    } catch (e) {
+      showSnackbar(e.message, 'error');
+    }
+  };
+
+  const ignore = async (id, boardName) => {
+    try {
+      const requestOptions = {
+        url: `https://rsclone-back-end.herokuapp.com/api/board/ignore/${id}`,
+        method: 'DELETE',
+        headers: { Authorization: `Bearer ${token}` },
+      };
+
+      await request(requestOptions);
+
+      await discardNotifications(id);
+      getNots(token, request);
+      showSnackbar(`You left the board ${boardName}!`, 'success');
+    } catch (e) {
+      showSnackbar(e.message, 'error');
+    }
+  };
+
+  useEffect(() => {
+    if (!isOpen) {
+      return;
+    }
+    getNots(token, request);
+  }, [isOpen, request, token, setNotification]);
+
+  const notices = nots.map((item) => {
+    const { from, boardName, boardId } = item;
+
+    const agree = () => {
+      comein(boardId, boardName);
+    };
+
+    const cancel = () => {
+      ignore(boardId, boardName);
+    };
+
+    return (
+      <NotificationCard
+        boardId={boardId}
+        boardName={boardName}
+        from={from}
+        agree={agree}
+        loading={loading}
+        cancel={cancel}
+      />
+    );
+  });
 
   return (
     <Collapse in={isOpen}>
@@ -42,7 +140,15 @@ const DropDown = ({ isOpen }) => {
         >
           Notifications
         </Typography>
-        <Notifications isOpen={isOpen} />
+        {nots.length > 0
+          ? notices
+          : (
+            <Typography
+              variant="body1"
+            >
+              You have no notifications right now
+            </Typography>
+          )}
       </div>
     </Collapse>
   );
